adjust pruning and finality duration
with increased blockrate syncing becomes quite slow and due to astrobwt our header verification takes more time to process, to compensate this we reduce the finality duration to 6h and the pruning duration to around 17.5h because we must maintain the div_ceil ratio of pruning_depth/finality_depth = 3. So with a 6-hour finality (172800 blocks) this requires a pruning duration between 345600 and 518400 blocks.